Updated poststart script to replace config files

+++ poststart.sh
@@ -7,23 +7,23 @@ if [ ! -d notebooks ]; then
     rm -rf /tmp/git-notebooks
 fi
 
-# set .condarc to configure conda envs
-# if [[ -d .condarc ]]; then
-#     mv .condarc .condarc.old
-# fi
-cp /config/.condarc /home/jovyan/
+# copy default run commands but provide a way for users to keep their own config
+if [ ! -f .keep-local-conf ]; then
+    # .condarc: set env_prompt, channels, envs_dirs,create_default_packages
+    cp /config/.condarc /home/jovyan/
+    # .Rprofile: set binary package repo
+    cp /config/.Rprofile /home/jovyan/
+    # .bash_profile: make login shells source .bashrc
+    echo "source ~/.bashrc" > .bash_profile
+    # .bashrc: activate myenv by default
+    echo "source activate myenv" > .bashrc
+fi
 
 # create default environment 'myenv'
 if [ ! -d my-conda-envs/myenv ]; then
     conda create --clone base --name myenv
     source activate myenv
 fi
-
-# set .Rprofile to use binary packages
-# if [[ -d .Rprofile ]]; then
-#     mv .Rprofile .Rprofile.old
-# fi
-cp /config/.Rprofile /home/jovyan/
 
 Rscript -e 'dir.create(path = Sys.getenv("R_LIBS_USER"), showWarnings = FALSE, recursive = TRUE)'
 Rscript -e '.libPaths( c( Sys.getenv("R_LIBS_USER"), .libPaths() ) )'
@@ -36,21 +36,9 @@ if [ ! -d /nfs ] || [ ! -d /lustre ] || [ ! -d /warehouse ]; then
     sudo mkdir -p /warehouse
 fi
 
-# make login shells source .bashrc
-if [ ! -d .bash_profile ]; then
-    echo "source ~/.bashrc" > .bash_profile
-fi
-
-if [ ! -d .bashrc ]; then
-    echo "source activate myenv" > .bash_profile
-fi
-
 # set env vars for nbresuse limits
 export MEM_LIMIT=$(cat /sys/fs/cgroup/memory/memory.limit_in_bytes)
 CPU_NANOLIMIT=$(cat /sys/fs/cgroup/cpu/cpu.cfs_quota_us)
 export CPU_LIMIT=$(($CPU_NANOLIMIT/100000))
 
 export USER=jovyan
-
-# download package information from all configured sources
-sudo apt-get update
